Fix layout and component block floating toolbars being shown behind other elements
