+++ .changeset/good-tigers-begin.md
@@ -1,0 +1,5 @@
+---
+'@keystone-6/fields-document': patch
+---
+
+Fixed layout and component block floating toolbars being shown behind other elements
