--- packages/fields-document/src/DocumentEditor/component-blocks/edit-mode.tsx
@@ -1,40 +1,0 @@
-import { Stack } from '@keystone-ui/core';
-import React, { useState } from 'react';
-import { Button as KeystoneUIButton } from '@keystone-ui/button';
-import { GenericPreviewProps } from './api';
-import {
-  FormValueContentFromPreviewProps,
-  NonChildFieldComponentSchema,
-} from './form-from-preview';
-
-export function FormValue({
-  onClose,
-  props,
-  isValid,
-}: {
-  props: GenericPreviewProps<NonChildFieldComponentSchema, unknown>;
-  onClose(): void;
-  isValid: boolean;
-}) {
-  const [forceValidation, setForceValidation] = useState(false);
-
-  return (
-    <Stack gap="xlarge" contentEditable={false}>
-      <FormValueContentFromPreviewProps {...props} forceValidation={forceValidation} />
-      <KeystoneUIButton
-        size="small"
-        tone="active"
-        weight="bold"
-        onClick={() => {
-          if (isValid) {
-            onClose();
-          } else {
-            setForceValidation(true);
-          }
-        }}
-      >
-        Done
-      </KeystoneUIButton>
-    </Stack>
-  );
-}
